Enabled preview and added functionality*
*WatchProviders
