+++ Shared/View/WatchProvidersList.swift
@@ -28,7 +28,7 @@ struct WatchProvidersList: View {
                                 if isConfirmationEnabled {
                                     showConfirmation.toggle()
                                 } else {
-                                    
+                                    openLink()
                                 }
                             }, label: {
                                 WatchProviderItem(item: item)
@@ -55,8 +55,14 @@ struct WatchProvidersList: View {
         }
     }
     
-    func openLink() {
-        
+    private func openLink() {
+        if let link = viewModel.link {
+#if os(macOS)
+            NSWorkspace.shared.open(link)
+#else
+            UIApplication.shared.open(link)
+#endif
+        }
     }
 }
 
@@ -86,143 +92,9 @@ private struct WatchProviderItem: View {
     }
 }
 
-//struct WatchProvidersList_Previews: PreviewProvider {
-//    static var previews: some View {
-//        WatchProvidersList()
-//    }
-//}
-
-class WatchProvidersListViewModel: ObservableObject {
-    @Published var isProvidersAvailable = false
-    @Published var items = [WatchProviderContent]()
-    @Published var link: URL?
-    private var isLoaded = false
-    @AppStorage("enableWatchProviders") private var isWatchProviderEnabled = true
-    @AppStorage("selectedWatchProviderRegion") private var watchRegion: WatchProviderOption = .us
-    @AppStorage("firstLocaleCheck") private var firstCheck = false
-    
-    private func checkLocale() {
-        if firstCheck { return }
-        let userLocale = Utilities.userRegion
-        let providerRegions = WatchProviderOption.allCases
-        for region in providerRegions {
-            if userLocale.lowercased() == region.rawValue.lowercased() {
-                watchRegion = region
-            }
-        }
-        firstCheck = true
-    }
-    
-    @MainActor
-    func load(id: ItemContent.ID, media: MediaType) async {
-        do {
-            if !firstCheck { checkLocale() }
-            if Task.isCancelled { return }
-            if isLoaded { return }
-            let providers = try await NetworkService.shared.fetchProviders(id: id, for: media)
-            if let results = providers.results {
-                var regionContent: ProviderItem?
-                switch watchRegion {
-                case .br:
-                    regionContent = results.br
-                case .us:
-                    regionContent = results.us
-                case .ae:
-                    regionContent = results.ae
-                case .ar:
-                    regionContent = results.ar
-                case .at:
-                    regionContent = results.at
-                case .au:
-                    regionContent = results.au
-                case .be:
-                    regionContent = results.be
-                case .bg:
-                    regionContent = results.bg
-                case .ca:
-                    regionContent = results.ca
-                case .ch:
-                    regionContent = results.ch
-                case .cz:
-                    regionContent = results.cz
-                case .de:
-                    regionContent = results.de
-                case .dk:
-                    regionContent = results.dk
-                case .ee:
-                    regionContent = results.ee
-                case .es:
-                    regionContent = results.es
-                case .fi:
-                    regionContent = results.fi
-                case .fr:
-                    regionContent = results.fr
-                case .gb:
-                    regionContent = results.gb
-                case .hk:
-                    regionContent = results.hk
-                case .hr:
-                    regionContent = results.hr
-                case .hu:
-                    regionContent = results.hu
-                case .id:
-                    regionContent = results.id
-                case .ie:
-                    regionContent = results.ie
-                case .india:
-                    regionContent = results.resultsIN
-                case .it:
-                    regionContent = results.it
-                case .jp:
-                    regionContent = results.jp
-                case .kr:
-                    regionContent = results.kr
-                case .lt:
-                    regionContent = results.lt
-                case .mx:
-                    regionContent = results.mx
-                case .nl:
-                    regionContent = results.nl
-                case .no:
-                    regionContent = results.no
-                case .nz:
-                    regionContent = results.nz
-                case .ph:
-                    regionContent = results.ph
-                case .pl:
-                    regionContent = results.pl
-                case .pt:
-                    regionContent = results.pt
-                case .rs:
-                    regionContent = results.rs
-                case .se:
-                    regionContent = results.se
-                case .sk:
-                    regionContent = results.sk
-                case .tr:
-                    regionContent = results.tr
-                case .za:
-                    regionContent = results.za
-                }
-                link = regionContent?.itemLink
-                var content = [WatchProviderContent]()
-                if let flatrate = regionContent?.flatrate {
-                    content.append(contentsOf: flatrate)
-                }
-                if let buy =  regionContent?.buy {
-                    content.append(contentsOf: buy)
-                }
-                if let free = regionContent?.free {
-                    content.append(contentsOf: free)
-                }
-                items.append(contentsOf: content.sorted { $0.listPriority < $1.listPriority })
-                withAnimation { isProvidersAvailable = true }
-            }
-            isLoaded = true
-        } catch {
-            let message = """
-"""
-            CronicaTelemetry.shared.handleMessage(message, for: "WatchProvidersListViewModel.load()")
-        }
+struct WatchProvidersList_Previews: PreviewProvider {
+    private static let example = ItemContent.previewContent
+    static var previews: some View {
+        WatchProvidersList(id: example.id, type: example.itemContentMedia)
     }
 }
